fix: prevent nodata-fill crash on integer bands
numpy.ma.MaskedArray.filled casts the fill value to the array's own
dtype, so filling an integer array with NaN raised TypeError. That path
is reachable via the model API whenever a float-declared band
(BandInfo.dtype defaults to float32) carries integer data. Cast to
float64 first; astype preserves the mask, so masked entries still
serialize as null.

+++ src/titiler_covjson/helpers.py
@@ -281,7 +281,12 @@ def numpy_dtype_to_ndarray(
     shape = list(data.shape)
 
     if covjson_dtype == "float":
-        floats = data.filled(np.nan).flatten().tolist()  # type: ignore[no-untyped-call, unused-ignore]
+        # Cast to float before filling: `dtype` is the band's *declared* dtype,
+        # which may be float over an integer array (BandInfo.dtype defaults to
+        # float32). MaskedArray.filled casts the fill value to the array's own
+        # dtype, so filling an int array with NaN would raise; astype preserves
+        # the mask. Masked entries become NaN, which pydantic serialises as null.
+        floats = data.astype(np.float64).filled(np.nan).flatten().tolist()  # type: ignore[no-untyped-call, attr-defined, unused-ignore]
         return NdArrayFloat(values=floats, axisNames=list(axis_names), shape=shape)
 
     # For int/str, build a boolean mask array (never scalar).
